Update index.js
I had a null reference exception "can not access revisions of null". The page name was: "Wy┼╝sza Szko┼éa Filologiczna" This was the only one that threw the error out of 30k wiki pages that I processed. I just added a check and returned an error message if that key happens to be null.

I also forgot to add a callback in my original PR at line 35, so I've included that as well. It would be a good idea to let the parse error trickle back down to the user in case they are using that callback function to iterate through a list, otherwise the program would halt.